fix: one gate contribution per finding, in every representation
Round 5 of review on the authoritative contract pipeline found the same
divergence one document over from the root-cause case already fixed: the
MCP `abi_compare` response publishes its own top-level `changes` array
beside the CLI report it embeds, and only the embedded one was zeroed for
findings the scoped `--used-by`/`--required-symbol` gate does not score.
Reproduced: `changes[].gate_contribution: 4` beside
`report.changes[].gate_contribution: 0` in one response, on a run that
exited 0.

The zeroing moves to `contract_gating.zero_scoped_out_gate_contributions`
so the two front ends share one implementation rather than each patching
its own array, and both call it before folding in the scoped-only and
missing-contract entries -- those are the scoped gate's own findings.

Also from review, each verified before fixing:

- the MCP scoped summary rebuild tallied `impact` (ChangeKind-derived) over
  every entry, so a NOT_EVALUATED finding could report `summary.breaking: 1`
  beside a verdict policy never scored it into. Not reachable through
  today's `abi_compare`, which takes no contract-domain selector -- the
  guard is what keeps that true by construction;
- the post-render loop stamped embedded missing-contract copies with the
  relevance-only helper while the top-level entry used the full-shape one.
  Both now use `stamp_missing_contract_entry` with the same resolved
  contribution;
- `_record_scoped_compatibility_decisions` now applies the same
  `is_evaluated` guard its pipeline counterpart does, so a future
  non-evaluated early-return kind cannot publish a decision beside
  `NOT_EVALUATED`;
- `contract_pipeline`'s `evaluated_for_policy` referenced a symbol
  `contract_gating` does not define; `checker_types` and `mcp_server` still
  described contract evaluation as shadow/advisory, which Phase 7 ended;
  `_promote_scoped_contract` took `Any` where the module already has the
  concrete types; the `_classified` id() set now documents the lifetime
  its safety rests on.

Tests: the CLI runner results in the new suite are checked before their
output is parsed, the two evidence-requiring domains are parametrized
rather than looped, and the new helpers carry annotations. Regressions
added for both the root-cause and the MCP representations.

Unrelated to the fixes, and the reason `post_processing.py` moves:
`post_processing.py` sat exactly at the 2000-line hard cap, so any edit to
it failed the AI-readiness gate. `PipelineContext` and the reachability
marker move to `post_processing_context.py` (157 lines) and
`post_processing_reachability.py` (533), leaving the parent at 1409.
`PipelineContext` is re-exported for its 23 importers. No new import cycle.
