+++ docs/reference/mcp-tools-reference.md
@@ -60,7 +60,7 @@ Compare two ABI surfaces and report breaking changes.
 | `used_by` | `list[str] \| None` | no | `None` | Application binary paths (ADR-043) — scope the comparison to what each app actually imports/requires, instead of the full library surface. old_input/new_input may be real library binaries or JSON snapshots carrying binary evidence (a dump of a real library, not headers-only). Mutually exclusive with required_symbols. Adds a ``used_by`` list to the response and computes ``exit_code`` from the worst-scoped app: the legacy BREAKING → 4 / API_BREAK → 2 floor, or — when any ``severity_*`` argument is given — the severity-aware scheme (which can return ``0`` for a scoped BREAKING verdict under ``severity_preset= "info-only"``). |
 | `required_symbols` | `list[str] \| None` | no | `None` | An explicit plugin/host required-entrypoint contract (ADR-043) — scope the comparison to only these exported symbols. Mutually exclusive with used_by. Adds a ``required_symbol_contract`` object to the response and computes ``exit_code`` from its verdict under the same legacy/severity-aware scheme as ``used_by`` above. |
 | `diagnostic_comparison` | `bool` | no | `False` | ADR-050 D2's sanctioned escape hatch. By default, a genuine ``ExtractionContract`` mismatch between old_input and new_input (old/new were not extracted under a comparable profile/scope) makes this tool return ``{"status": "not_comparable", "reason": ...}`` instead of a verdict. Setting this True downgrades that into a tentative diff whose report is stamped ``assurance: "none"``, so the caller can still see *a* result but knows not to trust it fully. Not needed, and does nothing, on a comparable pair. |
-| `contract_evaluation` | `bool` | no | `False` | ADR-049 Phase 3's shadow contract evaluator (non-authoritative). When True, each finding in the response's ``changes``/``out_of_surface_changes`` lists gains a ``contract_relevance`` (``IN_CONTRACT``/``PROVEN_OUT_OF_CONTRACT``/ ``UNKNOWN_UNPROVEN``/``UNKNOWN_UNRESOLVED``/``NOT_APPLICABLE``), ``contract_reason_code``, and — when resolved — ``contract_assurance`` field, reflecting whether the finding falls inside the library's declared public contract. Each finding also gains ``contract_evidence_refs`` (which evidence records its decision rests on), and — for ``output_format="json"`` only, since every other format returns a rendered report *string* — the response's ``report`` gains a ``contract_context`` block: the observed provider evidence, the resolved evaluation context, and the decision receipt, so a decision can be replayed or re-evaluated later without re-reading the binaries. The per-finding decisions are advisory: they never change ``verdict`` or which findings appear. **The orthogonal contract-coverage axis is not** (ADR-049 Phase 7): when the selected contract domain cannot be closed on the available evidence, that contributes ``1`` to ``exit_code``, so an otherwise compatible comparison can return ``exit_code: 1``. The contribution never lowers a real break's ``2``/``4``. The response's top-level ``contract_coverage`` block states what was found and what it contributed, whatever ``output_format`` is selected. Accepting incomplete coverage requires ``contract.unresolved: warn``, which only a ``kind: contract`` pack can state and which **this tool does not expose** — it is available on the ``compare`` and ``scan`` CLIs via ``--pack``. Default False; the response is unchanged from today's shape unless this is set. |
+| `contract_evaluation` | `bool` | no | `False` | ADR-049's contract-relevance evaluator. **Authoritative since Phase 7** — relevance is classified before compatibility policy, and policy scores only the findings it evaluates, so enabling this can change ``verdict``, ``compatibility_decision`` and ``exit_code``. When True, each finding in the response's ``changes``/``out_of_surface_changes`` lists gains a ``contract_relevance`` (``IN_CONTRACT``/``PROVEN_OUT_OF_CONTRACT``/ ``UNKNOWN_UNPROVEN``/``UNKNOWN_UNRESOLVED``/``NOT_APPLICABLE``), ``contract_reason_code``, and — when resolved — ``contract_assurance`` field, reflecting whether the finding falls inside the library's declared public contract. Each finding also gains ``contract_evidence_refs`` (which evidence records its decision rests on), and — for ``output_format="json"`` only, since every other format returns a rendered report *string* — the response's ``report`` gains a ``contract_context`` block: the observed provider evidence, the resolved evaluation context, and the decision receipt, so a decision can be replayed or re-evaluated later without re-reading the binaries. A finding this evaluator does not evaluate carries ``compatibility_evaluation_status: "NOT_EVALUATED"``, a ``null`` ``compatibility_decision`` and ``gate_contribution: 0`` — it stays fully reported and keeps its ``ChangeKind``, it just did not gate. **A second, orthogonal contract-coverage axis** (ADR-049 Phase 7): when the selected contract domain cannot be closed on the available evidence, that contributes ``1`` to ``exit_code``, so an otherwise compatible comparison can return ``exit_code: 1``. The contribution never lowers a real break's ``2``/``4``. The response's top-level ``contract_coverage`` block states what was found and what it contributed, whatever ``output_format`` is selected. Accepting incomplete coverage requires ``contract.unresolved: warn``, which only a ``kind: contract`` pack can state and which **this tool does not expose** — it is available on the ``compare`` and ``scan`` CLIs via ``--pack``. Default False; the response is unchanged from today's shape unless this is set. |
 
 ## `abi_deps`
